pyproject.toml
[tool.poetry]
name = "diff"
version = "0.1.0"
description = "Diffusion modeling ST data."
authors = ["romankouz <roko@umich.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
diffusers = "^0.31.0"
torch = "^2.5.1"
lightning = "^2.4.0"
matplotlib = "^3.9.2"
jax = "^0.4.35"
absl-py = "^2.1.0"
altair = "5.2.0"
altair-data-server = "0.4.1"
altair-saver = "0.5.0"
altair-viewer = "0.4.0"
arviz = "0.17.0"
asttokens = "2.4.1"
async-timeout = "4.0.3"
attrs = "23.2.0"
autograd = "1.6.2"
certifi = "2024.2.2"
chex = "0.1.85"
click = "8.1.7"
cloudpickle = "3.0.0"
comm = "0.2.1"
contextlib2 = "21.6.0"
contourpy = "^1.3.1"
cycler = "^0.12.1"
cython = "3.0.8"
dask = "2024.2.0"
debugpy = "1.8.1"
decorator = "5.1.1"
deneb = "1.0.1"
diffeqtorch = "1.0.0"
diffrax = "0.5.0"
distributed = "2024.2.0"
dm-tree = "^0.1.8"
entrypoints = "0.4"
equinox = "0.11.3"
exceptiongroup = "1.2.0"
executing = "2.0.1"
filelock = "3.13.1"
flatlatex = "0.15"
fonttools = "4.25.0"
fsspec = "2024.2.0"
future = "0.18.3"
gitdb = "4.0.11"
gitpython = "3.1.42"
gpy = "1.13.1"
greenlet = "3.0.3"
grpcio = "1.60.1"
h11 = "0.14.0"
h5netcdf = "1.3.0"
h5py = "3.10.0"
idna = "3.6"
importlib-metadata = "^8.5.0"
importlib-resources = "^6.4.5"
ipyparallel = "8.6.1"
ipython = "8.18.1"
jabbar = "0.0.16"
jaxlib = "^0.4.35"
jaxtyping = "0.2.25"
jedi = "0.19.1"
jinja2 = "3.1.3"
joblib = "1.3.2"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
julia = "0.6.2"
jupyter-client = "8.6.0"
jupyter-core = "5.7.1"
kiwisolver = "^1.4.7"
lineax = "0.0.4"
locket = "1.0.0"
markdown = "3.5.2"
markupsafe = "2.1.5"
matplotlib-inline = "0.1.6"
mergedeep = "1.3.4"
mkl-fft = "^1.3.11"
mkl-random = "^1.2.8"
mkl-service = "2.4.0"
ml-collections = "0.1.1"
mpmath = "1.3.0"
msgpack = "1.0.7"
munkres = "1.1.4"
nest-asyncio = "1.6.0"
networkx = "3.2.1"
nflows = "0.14"
numdifftools = "0.9.41"
opt-einsum = "^3.4.0"
optimistix = "0.0.6"
outcome = "1.3.0.post0"
packaging = "^24.2"
pandas = "2.2.0"
paramz = "0.9.6"
parso = "0.8.3"
partd = "1.4.1"
pexpect = "4.9.0"
pillow = "^11.0.0"
platformdirs = "4.2.0"
ply = "3.11"
portpicker = "1.6.0"
prompt-toolkit = "3.0.43"
protobuf = "4.25.3"
psutil = "5.9.8"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyabc = "0.12.13"
pyabcranger = "0.0.70"
pygments = "2.17.2"
pyknos = "0.15.2"
pyparsing = "^3.2.0"
pyqt5 = "5.15.10"
pyqt5-sip = "^12.15.0"
pyro-api = "0.1.2"
pyro-ppl = "1.8.6"
pysocks = "1.7.1"
python-dateutil = "^2.9.0.post0"
pytz = "2024.1"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
redis = "5.0.1"
referencing = "0.33.0"
regex = "2023.12.25"
rpds-py = "0.18.0"
sbi = "0.21.0"
sbibm = "1.1.0"
scikit-learn = "1.4.1.post1"
seaborn = "0.13.2"
selenium = "4.18.0"
sip = "^6.8.6"
six = "^1.16.0"
smmap = "5.0.1"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
sqlalchemy = "2.0.27"
stack-data = "0.6.3"
tblib = "3.0.0"
tensorboard = "2.16.2"
tensorboard-data-server = "0.7.2"
threadpoolctl = "3.3.0"
tomli = "^2.1.0"
toolz = "^1.0.0"
scanpy = "^1.10.4"
rpy2 = "^3.5.16"
gputil = "^1.4.0"
zuko = "^1.3.0"
torchvision = "^0.20.1"
omegaconf = "^2.3.0"
pyarrow = "^18.1.0"
igraph = "^0.11.8"
leidenalg = "^0.10.2"
louvain = "^0.8.2"
torch-geometric = "^2.6.1"
notebook = "^7.3.1"
pylibcugraphops-cu12 = "^24.12.0"
normflows = "^1.7.3"
phenograph = "^1.5.7"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
